Created a collectible prototype
To help test and for the artists to level design. : )

When the player (BP_Protagonist) comes within the radius of a pick-up, the pick-up destroys itself.

Keep in mind all collectibles will keep looking for the player to overlap, so it can be expensive (if this happens, we need to prototype it differently). : )