Benjamins stuff for functional BP prototype artists
* Updated .gitignore for 2 folders

Made the .gitignore account for the 2folders: 📂

Windows (this is the windows build folder)
.idea (Rider-specific, much like .sln for Visual Studio))

* Adding my own level

Generally for testing purposes - I did it by entering a level and going to "File" -> "Save As" -> and named it accordingly.

* Changing World Settings to my BPs

Had to change and adjust some settings to accommodate my testing environment. 😄

Changed World Settings (found under "Settings" top-right), and changed loaded Game Mode and default pawn to be my BP_Protagonist and GM_GameModeManager.

I did the same in Project Settings, under "Maps and Modes". These settings are overridden by any World Settings, fyi.

* Double Jump + Protagonist tweaks

Kinda as the title says, mostly featuring updates with double jump and the protagonist : )

* DefaultEditor.ini?

Not what exact kind of changes this is, but it is a lot of text inside the .ini file, so if this commit needs to be removed later, it's easily done.

* Created a collectible prototype

To help test and for the artists to level design. : )

When the player (BP_Protagonist) comes within the radius of a pick-up, the pick-up destroys itself.

Keep in mind all collectibles will keep looking for the player to overlap, so it can be expensive (if this happens, we need to prototype it differently). : )

* Updates to my level - adding collectibles

* Collectible no-Collision added

* Final adjustments my level