WorldObjectSettings via UDataAsset
It's a new class that allows us to create e.g. default variables that can be applied inside C++ classes.

Super useful and convenient, and it's like a BP kinda, but a lot tidier and better.

So far when you commit changes in these UDataAssets, or in the parent C++ class WorldObjectSettings, you have to restart Unreal to see those changes applied. Hope to remedy this later. 😄