Level Work
The work in progress for today, added more/brighter overall lighting, added a ton of decoration lighting that also helps the player see where to go in the stump level.

Moved the platforming down in the stump tutorial.
More lighting in the tutorial.

Changes to the main tree, changes in mesh and size.

Added a Mushroom bouncepad to the platforming on the secondfloor of the level. And moved the platforms a bit down.