Stage and AI Collision fixes (read descriptions)
For odd reasons, some of the areas in the game didn't have the right collision channels set. 🛑

I removed an unnecessary collider on the player, and realized something was iffy with the collisions.

This meant the player would clip through the initial tutorial stage, and AI wouldn't get the right overlap events going on (the player would be unaffected when colliding with AI). 👍

Should be fixed now. 👍