Pickup Overhaul
Pretty much just a big clean-up + overhaul and tidying of the premises, and including way more floating functionality now.

There's also the DefaultEngine.ini that contains redirects when I've felt the need to rename a function or UPROPERTY to a more fitting name; since it prevents issues down the line.